requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
